Switch to libuv queue work for async queries
The uv poll method before was showing serious issues
with timeouts/connects and consistent throughput.  This
new method now queues a 'synchronous' request on a libuv
background worker thread to run synchronously until
the result is complete or times out.  This simplifies the code a bit
at the cost of require a large number of background worker
threads to execute a lot of concurrent requests.

The other reason why this is desirable is because the
mysql_real_query() and mysql_store_result() functions were
hanging the event loop thread when libuv would trigger an on read
event.  This was causing the event loop to artificially reduce the
number of queries per second that they could execute as those
functions are synchronous and hang all other requests.  The
background worker threads avoid this problem since each query
is now isolated to its own worker thread.